refactor(headless): split Autocomplete into per-file parts for tree-shaking
