+++ packages/headless/src/primitives/autocomplete/parts.ts
@@ -1,0 +1,8 @@
+export { type AutocompleteProps, AutocompleteRoot as Root } from './autocomplete-root';
+export { type AutocompleteInputProps, AutocompleteInput as Input } from './autocomplete-input';
+export { type AutocompletePortalProps, AutocompletePortal as Portal } from './autocomplete-portal';
+export { type AutocompletePositionerProps, AutocompletePositioner as Positioner } from './autocomplete-positioner';
+export { type AutocompletePopupProps, AutocompletePopup as Popup } from './autocomplete-popup';
+export { type AutocompleteListProps, AutocompleteList as List } from './autocomplete-list';
+export { type AutocompleteOptionProps, AutocompleteOption as Option } from './autocomplete-option';
+export { type AutocompleteArrowProps, AutocompleteArrow as Arrow } from './autocomplete-arrow';
